+++ PRUNner/Backend/Utility/ObjectPools.cs
@@ -1,0 +1,9 @@
+using System.Text;
+
+namespace PRUNner.Backend.Utility
+{
+    public static class ObjectPools
+    {
+        public static readonly SimpleObjectPool<StringBuilder> StringBuilderPool = new(sb => sb.Clear());
+    }
+}